Completed day 9 (with all the refactoring put in).

+++ AdventOfCodeLib/Day02/Part1/Solution.cs
@@ -5,11 +5,11 @@ using System.Linq;
 
 namespace AdventOfCodeLib.Day02.Part1 {
 	public static class Solution {
-		public static int SolveFromProgram(string program, bool add1202 = true) {
+		public static long SolveFromProgram(string program, bool add1202 = true) {
 			return SolveFromProgramOutputProgram(program, add1202)[0];
 		}
-		public static List<int> SolveFromProgramOutputProgram(string program, bool add1202 = true) {
-			var convertedProgram = new List<int>(program.Split(",").Select(c => int.Parse(c)));
+		public static List<long> SolveFromProgramOutputProgram(string program, bool add1202 = true) {
+			var convertedProgram = new List<long>(program.Split(",").Select(c => long.Parse(c)));
 			if (add1202) {
 				convertedProgram[1] = 12;
 				convertedProgram[2] = 2;
@@ -17,12 +17,12 @@ namespace AdventOfCodeLib.Day02.Part1 {
 			return Solve(convertedProgram);
 		}
 
-		public static int SolveFromInputFile(string inputFile, bool add1202 = true) {
+		public static long SolveFromInputFile(string inputFile, bool add1202 = true) {
 			return SolveFromInputFileOutputProgram(inputFile, add1202)[0];
 		}
 
-		public static List<int> SolveFromInputFileOutputProgram(string inputFile, bool add1202 = true) {
-			var convertedProgram = new List<int>(File.ReadAllText(inputFile).Split(",").Select(c => int.Parse(c)));
+		public static List<long> SolveFromInputFileOutputProgram(string inputFile, bool add1202 = true) {
+			var convertedProgram = new List<long>(File.ReadAllText(inputFile).Split(",").Select(c => long.Parse(c)));
 			if (add1202) {
 				convertedProgram[1] = 12;
 				convertedProgram[2] = 2;
@@ -30,7 +30,7 @@ namespace AdventOfCodeLib.Day02.Part1 {
 			return Solve(convertedProgram);
 		}
 
-		private static List<int> Solve(List<int> program) {
+		private static List<long> Solve(List<long> program) {
 			var computer = new IntcodeComputer(program);
 			computer.RunProgram();
 			return computer.GetLastProgramState();
