+++ AdventOfCodeLib/Day05/Part2/Solution.cs
@@ -5,14 +5,14 @@ using System.Linq;
 
 namespace AdventOfCodeLib.Day05.Part2 {
 	public static class Solution {
-		public static int SolveFromInputFile(string inputFile) {
-			var convertedProgram = new List<int>(File.ReadAllText(inputFile).Split(",").Select(c => int.Parse(c)));
+		public static long SolveFromInputFile(string inputFile) {
+			var convertedProgram = new List<long>(File.ReadAllText(inputFile).Split(",").Select(c => long.Parse(c)));
 			return Solve(convertedProgram);
 		}
 
-		private static int Solve(List<int> program) {
-			var outputBuffer = new List<int>();
-			var computer = new IntcodeComputer(program, new Queue<int>(new int[] { 5 }), outputBuffer);
+		private static long Solve(List<long> program) {
+			var outputBuffer = new List<long>();
+			var computer = new IntcodeComputer(program, new Queue<long>(new long[] { 5 }), outputBuffer);
 			computer.RunProgram();
 			return outputBuffer.Last();
 		}
