+++ Tests/Day02/PartOneTests.cs
@@ -11,7 +11,7 @@ namespace Tests.Day02 {
 		[InlineData("2,4,4,5,99,0", "2,4,4,5,99,9801")]
 		[InlineData("1,1,1,4,99,5,6,0,99", "30,1,1,4,2,5,6,0,99")]
 		public void TestProgram(string program, string expectedProgram) {
-			Assert.Equal(new List<int>(expectedProgram.Split(",").Select(c => int.Parse(c))), Solution.SolveFromProgramOutputProgram(program, false));
+			Assert.Equal(new List<long>(expectedProgram.Split(",").Select(c => long.Parse(c))), Solution.SolveFromProgramOutputProgram(program, false));
 		}
 	}
 }
